fix: DTO 구조 변경에 따른 Controller/Service 컴파일 에러 수정 #90
- Controller/Service의 반환 타입을 구체적인 내부 클래스(CreateResponseDto 등)로 명시하여 타입 불일치 해결
- VerificationConverter의 빌더 패턴 사용 방식 수정 (Wrapper Class 적용)
- 인증 생성(Create) 및 조회(Read) 비즈니스 로직 통합 완료

+++ src/main/java/com/hrr/backend/domain/verification/service/VerificationServiceImpl.java
@@ -110,14 +110,14 @@ public class VerificationServiceImpl implements VerificationService {
         }
 
         LocalDate targetDate = targetDateTime.toLocalDate();
-        Integer certifiedCount = verificationRepository.countDistinctCertifiers(
+        Long certifiedCount = verificationRepository.countDistinctCertifiers(
                 currentRoundId,
                 VerificationStatus.COMPLETED,
                 targetDate.atStartOfDay(),
                 targetDate.atTime(LocalTime.MAX)
         );
 
-        return verificationConverter.toStatDto(certifiedCount, totalParticipantCount, targetDateTime);
+        return verificationConverter.toStatDto(certifiedCount.intValue(), totalParticipantCount, targetDateTime);
     }
 
     @Override
